refactor(frontend): break down the context into provider, context, utils and types
